YEL-74 [refactor] Vote Repository 추상화 작업
