+++ src/main/java/com/yello/server/domain/user/service/UserService.java
@@ -8,7 +8,7 @@ import com.yello.server.domain.user.dto.response.UserDetailResponse;
 import com.yello.server.domain.user.dto.response.UserResponse;
 import com.yello.server.domain.user.entity.User;
 import com.yello.server.domain.user.repository.UserRepository;
-import com.yello.server.domain.vote.entity.VoteRepository;
+import com.yello.server.domain.vote.repository.VoteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
