+++ src/main/java/com/yello/server/domain/vote/service/VoteService.java
@@ -1,29 +1,217 @@
 package com.yello.server.domain.vote.service;
 
+import static com.yello.server.domain.vote.common.WeightedRandomFactory.randomPoint;
+import static com.yello.server.global.common.ErrorCode.INVALID_VOTE_EXCEPTION;
+import static com.yello.server.global.common.ErrorCode.LACK_POINT_EXCEPTION;
+import static com.yello.server.global.common.ErrorCode.LACK_USER_EXCEPTION;
+import static com.yello.server.global.common.ErrorCode.NOT_FOUND_QUESTION_EXCEPTION;
+import static com.yello.server.global.common.factory.TimeFactory.minusTime;
+import static com.yello.server.global.common.util.ConstantUtil.COOL_DOWN_TIME;
+import static com.yello.server.global.common.util.ConstantUtil.KEYWORD_HINT_POINT;
+import static com.yello.server.global.common.util.ConstantUtil.NAME_HINT_DEFAULT;
+import static com.yello.server.global.common.util.ConstantUtil.NAME_HINT_POINT;
+import static com.yello.server.global.common.util.ConstantUtil.RANDOM_COUNT;
+import static com.yello.server.global.common.util.ConstantUtil.VOTE_COUNT;
+
+import com.yello.server.domain.cooldown.entity.Cooldown;
+import com.yello.server.domain.cooldown.entity.CooldownRepository;
+import com.yello.server.domain.friend.dto.response.FriendShuffleResponse;
+import com.yello.server.domain.friend.entity.Friend;
+import com.yello.server.domain.friend.entity.FriendRepository;
+import com.yello.server.domain.friend.exception.FriendException;
 import com.yello.server.domain.keyword.dto.response.KeywordCheckResponse;
+import com.yello.server.domain.keyword.entity.Keyword;
 import com.yello.server.domain.question.dto.response.QuestionForVoteResponse;
+import com.yello.server.domain.question.dto.response.QuestionVO;
+import com.yello.server.domain.question.entity.Question;
+import com.yello.server.domain.question.entity.QuestionRepository;
+import com.yello.server.domain.question.exception.QuestionException;
+import com.yello.server.domain.user.entity.User;
+import com.yello.server.domain.user.repository.UserRepository;
 import com.yello.server.domain.vote.dto.request.CreateVoteRequest;
-import com.yello.server.domain.vote.dto.response.*;
-import org.springframework.data.domain.Pageable;
-
+import com.yello.server.domain.vote.dto.response.RevealNameResponse;
+import com.yello.server.domain.vote.dto.response.VoteAvailableResponse;
+import com.yello.server.domain.vote.dto.response.VoteCreateResponse;
+import com.yello.server.domain.vote.dto.response.VoteDetailResponse;
+import com.yello.server.domain.vote.dto.response.VoteFriendResponse;
+import com.yello.server.domain.vote.dto.response.VoteListResponse;
+import com.yello.server.domain.vote.dto.response.VoteResponse;
+import com.yello.server.domain.vote.entity.Vote;
+import com.yello.server.domain.vote.exception.VoteForbiddenException;
+import com.yello.server.domain.vote.exception.VoteNotFoundException;
+import com.yello.server.domain.vote.repository.VoteRepository;
+import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface VoteService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class VoteService {
 
-    VoteListResponse findAllVotes(Long userId, Pageable pageable);
+    private final UserRepository userRepository;
+    private final FriendRepository friendRepository;
+    private final QuestionRepository questionRepository;
+    private final CooldownRepository cooldownRepository;
+    private final VoteRepository voteRepository;
 
-    VoteDetailResponse findVoteById(Long id);
+    public VoteListResponse findAllVotes(Long userId, Pageable pageable) {
+        Integer count = voteRepository.countAllByReceiverUserId(userId);
+        List<VoteResponse> votes = voteRepository.findAllByReceiverUserId(userId, pageable)
+            .stream()
+            .map(VoteResponse::of)
+            .toList();
+        return VoteListResponse.of(count, votes);
+    }
 
-    List<VoteFriendResponse> findAllFriendVotes(Long userId, Pageable pageable);
+    @Transactional
+    public VoteDetailResponse findVoteById(Long id) {
+        Vote vote = voteRepository.findById(id);
+        vote.read();
+        return VoteDetailResponse.of(vote);
+    }
 
-    KeywordCheckResponse checkKeyword(Long userId, Long voteId);
+    public List<VoteFriendResponse> findAllFriendVotes(Long userId, Pageable pageable) {
+        return voteRepository.findAllReceivedByFriends(userId, pageable)
+            .stream()
+            .map(VoteFriendResponse::of)
+            .toList();
+    }
 
-    List<QuestionForVoteResponse> findVoteQuestionList(Long userId);
+    @Transactional
+    public KeywordCheckResponse checkKeyword(Long userId, Long voteId) {
+        Vote vote = voteRepository.findById(voteId);
+        User user = userRepository.findById(userId);
 
-    VoteAvailableResponse checkVoteAvailable(Long userId);
+        vote.checkKeyword();
 
-    VoteCreateResponse createVote(Long userId, CreateVoteRequest request);
+        if (user.getPoint() < KEYWORD_HINT_POINT) {
+            throw new VoteForbiddenException(LACK_POINT_EXCEPTION);
+        }
 
-    RevealNameResponse revealNameHint(Long userId, Long voteId);
+        user.minusPoint(KEYWORD_HINT_POINT);
+        return KeywordCheckResponse.of(vote);
+    }
+
+    public List<QuestionForVoteResponse> findVoteQuestionList(Long userId) {
+        User user = userRepository.findById(userId);
+
+        List<Friend> friends = friendRepository.findAllByUserId(user.getId());
+        if (friends.size() < RANDOM_COUNT) {
+            throw new FriendException(LACK_USER_EXCEPTION);
+        }
+
+        List<Question> questions = questionRepository.findAll();
+        Collections.shuffle(questions);
+
+        List<Question> questionList = questions.stream()
+            .limit(VOTE_COUNT)
+            .toList();
+
+        return questionList.stream()
+            .map(question -> generateVoteQuestion(user, question))
+            .toList();
+    }
+
+    public VoteAvailableResponse checkVoteAvailable(Long userId) {
+        User user = userRepository.findById(userId);
+        List<Friend> friends = friendRepository.findAllByUserId(user.getId());
+
+        if (friends.size() < RANDOM_COUNT) {
+            throw new FriendException(LACK_USER_EXCEPTION);
+        }
+
+        Cooldown cooldown = cooldownRepository.findByUserId(user.getId())
+            .orElse(Cooldown.of(user, minusTime(LocalDateTime.now(), COOL_DOWN_TIME)));
+
+        return VoteAvailableResponse.of(user, cooldown);
+    }
+
+    @Transactional
+    public VoteCreateResponse createVote(Long userId, CreateVoteRequest request) {
+        User sender = userRepository.findById(userId);
+
+        request.voteAnswerList()
+            .forEach(voteAnswer -> {
+                User receiver = userRepository.findById(voteAnswer.friendId());
+                Question question = findQuestion(voteAnswer.questionId());
+
+                Vote newVote = Vote.createVote(voteAnswer.keywordName(), sender, receiver, question,
+                    voteAnswer.colorIndex());
+                voteRepository.save(newVote);
+            });
+
+        Optional<Cooldown> cooldown = cooldownRepository.findByUserId(sender.getId());
+        if (cooldown.isEmpty()) {
+            cooldownRepository.save(Cooldown.of(sender, LocalDateTime.now()));
+        } else {
+            cooldown.get().updateDate(LocalDateTime.now());
+        }
+
+        sender.plusPoint(request.totalPoint());
+        return VoteCreateResponse.of(sender.getPoint());
+    }
+
+    @Transactional
+    public RevealNameResponse revealNameHint(Long userId, Long voteId) {
+        User sender = userRepository.findById(userId);
+
+        if (sender.getPoint() < NAME_HINT_POINT) {
+            throw new VoteForbiddenException(LACK_POINT_EXCEPTION);
+        }
+
+        Vote vote = voteRepository.findById(voteId);
+        if (vote.getNameHint()!=NAME_HINT_DEFAULT) {
+            throw new VoteNotFoundException(INVALID_VOTE_EXCEPTION);
+        }
+
+        int randomIndex = (int) (Math.random() * 2);
+        vote.checkNameIndexOf(randomIndex);
+        sender.minusPoint(NAME_HINT_POINT);
+
+        return RevealNameResponse.of(vote.getSender(), randomIndex);
+    }
+
+    private QuestionForVoteResponse generateVoteQuestion(User user, Question question) {
+        List<Keyword> keywordList = question.getKeywordList();
+        Collections.shuffle(keywordList);
+
+        return QuestionForVoteResponse.builder()
+            .friendList(getShuffledFriends(user))
+            .keywordList(getShuffledKeywords(question))
+            .question(QuestionVO.of(question))
+            .questionPoint(randomPoint())
+            .build();
+    }
+
+    private List<FriendShuffleResponse> getShuffledFriends(User user) {
+        List<Friend> allFriend = friendRepository.findAllByUserId(user.getId());
+        Collections.shuffle(allFriend);
+
+        return allFriend.stream()
+            .map(FriendShuffleResponse::of)
+            .limit(RANDOM_COUNT)
+            .toList();
+    }
+
+    private List<String> getShuffledKeywords(Question question) {
+        List<Keyword> keywordList = question.getKeywordList();
+        Collections.shuffle(keywordList);
+
+        return keywordList.stream()
+            .map(Keyword::getKeywordName)
+            .limit(RANDOM_COUNT)
+            .toList();
+    }
+
+    private Question findQuestion(Long questionId) {
+        return questionRepository.findById(questionId)
+            .orElseThrow(() -> new QuestionException(NOT_FOUND_QUESTION_EXCEPTION));
+    }
 
 }
